+++ jest.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  transform: {
+    '^.+\\.(js|jsx|ts|tsx)$': 'babel-jest',
+  },
+  transformIgnorePatterns: [
+    '/node_modules/(?!cheerio|parse5|domutils|dom-serializer|htmlparser2|entities|css-select|css-what|domelementtype|domhandler|nth-check|boolbase)'
+  ],
+  moduleFileExtensions: ['js', 'jsx', 'ts', 'tsx', 'json', 'node'],
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: [
+    '<rootDir>/src/setupTests.ts',
+    '<rootDir>/api/setupTests.js'
+  ],
+};